app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-storage-blob==12.19.0
asyncpg==0.29.0
sqlalchemy==2.0.23
python-dotenv==1.0.0